+++ LibTest/core/int/operator_addition_A01_t01.dart
@@ -22,9 +22,10 @@ main() {
   Expect.equals(-8301034833169298229, 9223372036854775807 + 922337203685477580);
   Expect.equals(-1, -9223372036854775808 + 9223372036854775807);
 
-  final double MIN_DOUBLE = Math.pow(2.0, -1074);
+  final double MIN_DOUBLE = Math.pow(2.0, -1074) as double;
   final double NEG_MIN_DOUBLE = -1 * MIN_DOUBLE; 
-  final double MAX_DOUBLE = (2 - Math.pow(2.0, -52)) * Math.pow(2.0, 1023);
+  final double MAX_DOUBLE =
+      (2 - Math.pow(2.0, -52)) * Math.pow(2.0, 1023) as double;
   final double NEG_MAX_DOUBLE = -1 * MAX_DOUBLE; 
 
   Expect.equals(0.0, 0 + 0.0);
